switching target of the enemy
It's not working but if there's someone else other than Naimo and I who could look at it it would be nice. It's probably really easy but our brains are worked up and we're officially dead. This is zoombie Nida writing now bye enjoy the last days of your lives.

Co-Authored-By: naimoibrahim75 <naimoibrahim75@users.noreply.github.com>